mock remove from user answers
