+++ test/controllers/subscription/ManageGroupDetailCheckYourAnswersControllerSpec.scala
@@ -17,14 +17,23 @@
 package controllers.subscription
 
 import base.SpecBase
-import models.subscription.AccountingPeriod
-import models.{MneOrDomestic, UserAnswers}
+import connectors.UserAnswersConnectors
+import models.subscription.{AccountingPeriod, AmendSubscriptionRequestParameters, DashboardInfo}
+import models.{ApiError, MneOrDomestic, NonUKAddress, SubscriptionCreateError, UserAnswers}
+import org.mockito.ArgumentMatchers.any
+import org.mockito.Mockito.when
 import pages._
+import play.api.inject
+import play.api.inject.bind
+import play.api.libs.json.{JsValue, Json}
 import play.api.test.FakeRequest
 import play.api.test.Helpers._
+import services.AmendSubscriptionService
+import uk.gov.hmrc.http.{HeaderCarrier, HttpResponse}
 import viewmodels.govuk.SummaryListFluency
 
 import java.time.LocalDate
+import scala.concurrent.Future
 
 class ManageGroupDetailCheckYourAnswersControllerSpec extends SpecBase with SummaryListFluency {
 
@@ -34,7 +43,25 @@ class ManageGroupDetailCheckYourAnswersControllerSpec extends SpecBase with Summ
     val endDate   = LocalDate.of(2025, 12, 31)
     val date      = AccountingPeriod(startDate, endDate)
 
-    "must return OK and the correct view if an answer is provided to every question " in {
+    val amendSubUserAnswers = emptyUserAnswers
+      .setOrException(upeRegisteredInUKPage, true)
+      .setOrException(upeNameRegistrationPage, "International Organisation Inc.")
+      .setOrException(subPrimaryContactNamePage, "Name")
+      .setOrException(subPrimaryEmailPage, "email@email.com")
+      .setOrException(subPrimaryPhonePreferencePage, true)
+      .setOrException(subPrimaryCapturePhonePage, "123456789")
+      .setOrException(subAddSecondaryContactPage, true)
+      .setOrException(subSecondaryContactNamePage, "second contact name")
+      .setOrException(subSecondaryEmailPage, "second@email.com")
+      .setOrException(subSecondaryPhonePreferencePage, true)
+      .setOrException(subSecondaryCapturePhonePage, "123456789")
+      .setOrException(subRegisteredAddressPage, NonUKAddress("this", None, "over", None, None, countryCode = "AR"))
+      .setOrException(subMneOrDomesticPage, MneOrDomestic.Uk)
+      .setOrException(subAccountingPeriodPage, date)
+      .setOrException(fmDashboardPage, DashboardInfo("org name", LocalDate.of(2025, 12, 31)))
+      .setOrException(NominateFilingMemberPage, false)
+
+    "return OK and the correct view if an answer is provided to every question " in {
       val userAnswer = UserAnswers(userAnswersId)
         .set(subMneOrDomesticPage, MneOrDomestic.Uk)
         .success
@@ -52,7 +79,7 @@ class ManageGroupDetailCheckYourAnswersControllerSpec extends SpecBase with Summ
       }
     }
 
-    "must return OK and the correct view if an answer is provided to every question when UkAndOther  option is selected  " in {
+    "return OK and the correct view if an answer is provided to every question when UkAndOther  option is selected  " in {
       val userAnswer = UserAnswers(userAnswersId)
         .set(subMneOrDomesticPage, MneOrDomestic.UkAndOther)
         .success
@@ -69,5 +96,42 @@ class ManageGroupDetailCheckYourAnswersControllerSpec extends SpecBase with Summ
         contentAsString(result) must include("Where does the group operate?")
       }
     }
+
+    "trigger amend subscription API if all data is available for group accounting period" in {
+      val application = applicationBuilder(userAnswers = Some(amendSubUserAnswers))
+        .overrides(inject.bind[AmendSubscriptionService].toInstance(mockAmendSubscriptionService))
+        .overrides(bind[UserAnswersConnectors].toInstance(mockUserAnswersConnectors))
+        .build()
+      running(application) {
+        val mockHttpResponse = HttpResponse(OK, "")
+        when(mockUserAnswersConnectors.remove(any())(any())).thenReturn(Future.successful(mockHttpResponse))
+        val mockResponse = Json.parse("""{"success":{"processingDate":"2022-01-31T09:26:17Z","formBundleNumber":"119000004320"}}""")
+        when(mockAmendSubscriptionService.amendSubscription(AmendSubscriptionRequestParameters(any()))(any()))
+          .thenReturn(Future.successful(Right(mockResponse)))
+
+        val request = FakeRequest(POST, controllers.subscription.manageAccount.routes.ManageGroupDetailsCheckYourAnswersController.onSubmit.url)
+        val result  = route(application, request).value
+        status(result) mustBe SEE_OTHER
+        redirectLocation(result) mustBe Some("/report-pillar2-top-up-taxes/pillar2-top-up-tax-home")
+      }
+    }
+
+    "redirect to journey recovery if no data if no data is found for group accounting period" in {
+      val mockAmendSubscriptionService = mock[AmendSubscriptionService]
+      val application = applicationBuilder(userAnswers = None)
+        .overrides(inject.bind[AmendSubscriptionService].toInstance(mockAmendSubscriptionService))
+        .build()
+
+      val mockResponse: Either[ApiError, JsValue] = Left(SubscriptionCreateError)
+      when(mockAmendSubscriptionService.amendSubscription(any[AmendSubscriptionRequestParameters])(any[HeaderCarrier]))
+        .thenReturn(Future.successful(mockResponse))
+
+      val request = FakeRequest(POST, controllers.subscription.manageAccount.routes.ManageGroupDetailsCheckYourAnswersController.onSubmit.url)
+      val result  = route(application, request).value
+
+      status(result) mustBe SEE_OTHER
+      redirectLocation(result) mustBe Some(controllers.routes.JourneyRecoveryController.onPageLoad().url)
+    }
+
   }
 }
